Move StreamDecoder factories
